+++ Backend/HTTP_Testing/Login_test.http
@@ -2,6 +2,6 @@ POST http://localhost:3200/users/login
 content-type: application/json
 
 {
-  "Email":"testing",
-  "Password":"asdlfjasdkjfsd"
+  "Email":"coltoniscool",
+  "Password":"password"
 }
