+++ Frontend/finance_manager/src/app/api_urls/URL.ts
@@ -2,6 +2,6 @@ export const endpoints =
 {
   'checkLogin':'http://localhost:3200/users/login',
   'signup':'http://localhost:3200/users/signup',
-  'getUserDetails':'http://localhost:3200/users/finance/',//append `${this.email for proper usage}`
+  'getUserDetails':'http://localhost:3200/form/getForm/',//append `${this.email for proper usage}`
   'postForm': 'http://localhost:3200/form/postDetails'
 };
